third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/dedicated-worker-cache-storage.https-expected.txt

This is a testharness.js-based test.
PASS coep-none coep-none corp-cross-origin
PASS coep-none coep-none corp-undefined
PASS coep-none coep-require-corp corp-cross-origin
PASS coep-none coep-require-corp corp-undefined
PASS coep-require-corp coep-none corp-cross-origin
FAIL coep-require-corp coep-none corp-undefined assert_equals: expected false but got true
PASS coep-require-corp coep-require-corp corp-cross-origin
FAIL coep-require-corp coep-require-corp corp-undefined assert_equals: expected false but got true
Harness: the test ran to completion.
